+++ Extensions/IISWebAppDeploy/Src/CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Support of wild card input file pattern for sql deployment task.
+
 ## [1.4.2] - 4th January 2017
 ### Fixed
 - Assign SSL certificate to HTTPS binding when passing IP and port